+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ###### Changelog
 
+### 0.9.9 beta 1:
+- Corregido bug en canal LiveStream.
+- Añadido canal de series Seriesflv.net
+
 
 ### 0.9.8